pyproject.toml
[project]
name = "racebot"
version = "0.0.0"
description = "Discord countdown bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.6.4",
]

[tool.ruff]
line-length = 120
lint.select = ["ALL"]
lint.ignore = [
    "A002",
    "EM101",
    "EM102",
    "FIX002",
    "S101",
    "T201",
    "TRY003",
    "TRY300",
    "COM812",
    "RUF002",
    "D",
    "TD",
]
lint.fixable = ["ALL"]

[tool.ruff.format]
docstring-code-format = true
